:snake: :stethoscope: :package: add observability healthcheck module
